ajax+rest

+++ src/main/java/web/dao/UserDao.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface UserDao {
     void addUser(User application);
-    void deleteUser(int userId);
+    void deleteUser(Long userId);
     void updateUser(User application);
     List<User> getAllUsers();
-    User getUserById(int userId);
+    User getUserById(Long userId);
     User getUserByUsername(String username);
 }
